feat: add fronting domain option
Adds an option for custom fronting domain. The default fronting method
is changed to directly testing with speed.cldouflare.com.
If the default fronting address is blocked, the user should create an
dns record (CNAME) to speed.cloudflare.com and pass it using the added
option.

Resolves: #629, #626
