--- python/cfscanner/__init__.py
@@ -1,4 +1,0 @@
-from .main import main
-
-if __name__ == "__main__":
-    main()
