--- python/cfscanner/report/__init__.py
@@ -1,1 +1,0 @@
-from .colors import Colors
