--- python/cfscanner/report/colors.py
@@ -1,9 +1,0 @@
-class Colors:
-    OKBLUE = '\033[94m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    ARGHDR = '\033[95m'
-    
-
