--- python/cfscanner/subnets/__init__.py
@@ -1,1 +1,0 @@
-from .cidr import read_cidrs, read_cidrs_from_file, read_cidrs_from_url, get_num_ips_in_cidr, cidr_to_ip_gen
