--- python/cfscanner/subnets/regex.py
@@ -1,1 +1,0 @@
-CIDR_V4_REGEX = "(?:\d{1,3}\.){3}\d{1,3}\/\d{1,2}"
