--- python/cfscanner/xray/__init__.py
@@ -1,23 +1,0 @@
-SUPPORTED = [
-    ('android', 'arm64', 'v8a'),
-    ('linux', 'arm32', 'v5'),
-    ('linux', 'arm32', 'v6'),
-    ('linux', 'arm32', 'v7a'),
-    ('linux', 'arm64', 'v8a'),
-    ('linux', 'mips32le'),
-    ('linux', 'mips32'),
-    ('linux', 'ppc64le'),
-    ('linux', 'ppc64'),
-    ('linux', 's390x'),
-    ('linux', 'riscv64'),
-    ('linux', '64'),
-    ('linux', '32'),
-    ('windows', 'arm64', 'v8a'),
-    ('windows', 'arm32', 'v7a'),
-    ('windows', '64'),
-    ('windows', '32'),
-    ('macos', 'arm64', 'v8a'),
-    ('macos', '64')
-]
-
-LATEST_SUPPORTED_VERSION = "1.8.7"
