+++ python/src/cfscanner/subnets/regex.py
@@ -1,0 +1,1 @@
+CIDR_V4_REGEX = "(?:\d{1,3}\.){3}\d{1,3}\/\d{1,2}"
